+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fix moving Conditional/input code during upgrade
  - Generate Conditional/input code for next upgrade
  - Fix MQTT Input ([#685](https://github.com/kizniche/mycodo/issues/685))
+ - Fix Atlas pH Input issue ([#686](https://github.com/kizniche/mycodo/issues/686))
 
 
 ## 7.7.1 (2019.09.08)